+++ imagesmacker/fonts.py
@@ -16,7 +16,7 @@ def font_loader(
 class FontSizeCalculator:
     def __init__(
         self,
-        draw: ImageDraw.ImageDraw,
+        img_draw: ImageDraw.ImageDraw,
         font_path: str,
     ) -> None:
         """
@@ -28,12 +28,17 @@ class FontSizeCalculator:
         - kwargs (`dict[str, Any]`): Additional keyword arguments.
 
         """
-        self.draw = draw
+        self.img_draw = img_draw
         self.font_path = font_path
 
-    def get_text_bbox(self, size: int, text: str) -> tuple[int, int]:
+    def get_multiline_baseline_bbox(self, size: int, text: str) -> tuple[int, int]:
         """
-        Get the width and height of the text for a given font size.
+        Returns a baseline-aware bounding box for multiline text.
+
+        Baseline of the FIRST line is at y = 0.
+
+        top = -ascent_of_first_line
+        bottom = (sum of all line heights) + spacing*(n_lines-1) - ascent_of_first_line
 
         Args:
         - size (`int`): The font size to measure.
@@ -43,9 +48,27 @@ class FontSizeCalculator:
         `list[int]`: A list containing the width and height of the text.
 
         """
-        x1, y1, x2, y2 = self.draw.multiline_textbbox(
-            xy=(0, 0),
-            text=text,
-            font=font_loader(self.font_path, size),
-        )
-        return (round(x2 - x1), round(y2 - y1))
+
+        font = font_loader(self.font_path, size)
+
+        lines = text.split("\n")
+        ascent, descent = font.getmetrics()
+
+        # Width = max width of any line
+        widths = [font.getsize(line)[0] for line in lines]
+
+        # x1, y1, x2, y2 = self.img_draw.multiline_textbbox(
+        #     xy=(0, 0),
+        #     text=text,
+        #     font=font_loader(self.font_path, size),
+        # )
+
+        max_width = max(widths)
+
+        # Height per line = ascent + descent
+        line_height = ascent + descent
+
+        # Total height
+        total_height = (line_height * len(lines)) + (spacing * (len(lines) - 1))
+
+        return (max_width, total_height - ascent)